+++ examples/with-one-login-many-subdomains/frontend/src/Home/CallAPIView.js
@@ -1,7 +1,7 @@
 import axios from "axios";
 import Session from "supertokens-auth-react/recipe/session";
 import { useHistory } from "react-router-dom";
-import { getApiDomain } from "../App";
+import { getApiDomain } from "../utils";
 Session.addAxiosInterceptors(axios);
 
 export default function CallAPIView() {
